// lib/api/generated/models/userAuthResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ServiceDesk API
 * OpenAPI spec version: 1.0
 */
import type { SocialNetworks } from './socialNetworks';

export interface UserAuthResponse {
  active?: boolean;
  avatarUrl?: string;
  departmentName?: string;
  fio?: string;
  id?: number;
  permissions?: string[];
  positionName?: string;
  roles?: string[];
  socialNetwork?: SocialNetworks;
  specialist?: boolean;
  specialistType?: string;
  username?: string;
}